+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__/
 uploads/
 chroma_db_multimodal/
 static/
+.env
